+++ src/server/views/dashboard/queueJobsByState.js
@@ -80,6 +80,7 @@ async function _html(req, res) {
 
   const page = parseInt(req.query.page, 10) || 1;
   const pageSize = parseInt(req.query.pageSize, 10) || 100;
+  const order = req.query.order || 'desc';
 
   const startId = (page - 1) * pageSize;
   const endId = startId + pageSize - 1;
@@ -100,8 +101,10 @@ async function _html(req, res) {
     // Filter out Bee jobs that have already been removed by the time the promise resolves
     jobs = jobs.filter((job) => job);
   } else {
-    jobs = await queue[`get${_.capitalize(state)}`](startId, endId);
+    const stateTypes = state === 'waiting' ? ['wait', 'paused'] : state;
+    jobs = await queue.getJobs(stateTypes, startId, endId, order === 'asc');
     jobs = jobs.filter((job) => job && typeof job === 'object') // Filter out Bull jobs that have become null by the time the promise resolves
+    /*
     // Wait for all logs to be added to jobs before rendering page
     jobs = await Promise.all(
         jobs.map(async (job) => {
@@ -110,6 +113,7 @@ async function _html(req, res) {
           return job;
         })
     );
+    */
   }
 
   for (const job of jobs) {
@@ -123,6 +127,7 @@ async function _html(req, res) {
     pages.push(_.last(pages) + 1);
   }
   pages = pages.filter((page) => page <= _.ceil(jobCounts[state] / pageSize));
+  const disableRetry = !(state === 'failed' || (state === 'delayed' && !queue.IS_BEE));
 
   return res.render('dashboard/templates/queueJobsByState', {
     basePath,
@@ -132,10 +137,13 @@ async function _html(req, res) {
     jobs,
     jobsInStateCount: jobCounts[state],
     disablePagination: queue.IS_BEE && (state === 'succeeded' || state === 'failed'),
+    disableOrdering: queue.IS_BEE,
+    disableRetry,
     currentPage: page,
     pages,
     pageSize,
     lastPage: _.last(pages),
+    order,
   });
 }